Create trait for ThermalStorageCalculations
