+++ src/main/scala/edu/ie3/simona/model/thermal/ThermalStorageCalculations.scala
@@ -1,0 +1,61 @@
+/*
+ * © 2024. TU Dortmund University,
+ * Institute of Energy Systems, Energy Efficiency and Energy Economics,
+ * Research group Distribution grid planning and operation
+ */
+
+package edu.ie3.simona.model.thermal
+
+import edu.ie3.util.scala.quantities.SpecificHeatCapacity
+import edu.ie3.util.scala.quantities.SquantsUtils.RichEnergy
+import squants.space.Volume
+import squants.{Energy, Temperature}
+
+trait ThermalStorageCalculations {
+
+  /** Equation from docs for the relation between needed volume and energy.
+    *
+    * @param volume
+    *   needed/available volume
+    * @param c
+    *   Specific heat capacity
+    * @param inletTemp
+    *   Inlet temperature
+    * @param returnTemp
+    *   Return temperature
+    * @return
+    *   energy
+    */
+  def volumeToEnergy(
+      volume: Volume,
+      c: SpecificHeatCapacity,
+      inletTemp: Temperature,
+      returnTemp: Temperature,
+  ): Energy = {
+    c.calcEnergy(returnTemp, inletTemp, volume)
+  }
+
+  /** Equation from docs for the relation between stored heat and volume change.
+    *
+    * @param energy
+    *   available energy
+    * @param c
+    *   Specific heat capacity
+    * @param inletTemp
+    *   Inlet temperature
+    * @param returnTemp
+    *   Return temperature
+    * @return
+    *   volume
+    */
+  def energyToVolume(
+      energy: Energy,
+      c: SpecificHeatCapacity,
+      inletTemp: Temperature,
+      returnTemp: Temperature,
+  ): Volume = {
+    val energyDensity = c.calcEnergyDensity(returnTemp, inletTemp)
+
+    energy.calcVolume(energyDensity)
+  }
+}
